backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Talisman==1.0.0
python-dotenv==1.0.1
openai==1.33.0
pandas==2.2.2
numpy==1.26.4